Putting signal traps for better cleanup detection...
The issue of a setup function failing and then leaving the system in a
strange state is what is being addressed here.
The expectation is that for every "setup_" function there is a
companion "cleanup_" function.  Upon an INT or TERM signal the cleanup
function is called.

(TODO ... for times when there is no "cleanup_" function try to do the
right thing and not just barf)